requirements.txt
streamlit
pandas<2.1.0
beautifulsoup4
requests
python-dotenv
geopy
firebase-admin
